+++ client/src/components/modals/Modal.tsx
@@ -1,59 +1,66 @@
-
-import tw from "twin.macro";
-import { MdOutlineClose } from "react-icons/md";
+import tw from 'twin.macro';
+import { MdOutlineClose } from 'react-icons/md';
 
 import Button from '../buttons/Button';
-import { useDispatch } from "react-redux";
-import { modalActions } from "../../store/modalSlice";
-import { ModalType } from "../type";
+import { useDispatch } from 'react-redux';
+import { modalActions } from '../../store/modalSlice';
+import { ModalType } from '../../types';
 
 type ModalProps = {
-    type?: ModalType, 
-}
-const Modal = ({type}: ModalProps) => {
-    const title:Array<string> = ["후즈북의 큐레이터가 되신것을 환영합니다!", "OO 님의 큐레이션 구독을 취소하시겠어요?"];
-    const dispatch = useDispatch();
-    const handleCloseModal = () => {
-        dispatch(modalActions.close());
-    }
-    
-    //구독 취소 요청 보내기
-    // const handleSubscribe = () => {
+  type?: ModalType;
+};
+const Modal = ({ type }: ModalProps) => {
+  const title: Array<string> = [
+    '후즈북의 큐레이터가 되신것을 환영합니다!',
+    'OO 님의 큐레이션 구독을 취소하시겠어요?',
+  ];
+  const dispatch = useDispatch();
+  const handleCloseModal = () => {
+    dispatch(modalActions.close());
+  };
 
-    // }
+  //구독 취소 요청 보내기
+  // const handleSubscribe = () => {
 
-    return(
-        <ModalBackdrop>
-            <ModalView>
-                <CloseBtn onClick={handleCloseModal}>
-                    <MdOutlineClose size="1.2rem"/>
-                </CloseBtn>
-                
-                
-                { type === ModalType.WELCOME ?
-                (
-                    <>
-                    <ModalTitle>{title[0]}</ModalTitle>
-                    <ButtonZone>
-                        <Button type="primary" content="반가워요" onClick={handleCloseModal}/>
-                    </ButtonZone>
-                    </>
-                ) : (
-                    <>
+  // }
 
-                    <ModalTitle>{title[1]}</ModalTitle>
-                    <ButtonZone>
-                        <Button type="cancel" content="구독취소"  onClick={handleCloseModal} width="calc(30%-0.5rem)"/>
-                        <Button type="basic" content="닫기" onClick={handleCloseModal} width="calc(40%-0.5rem)"/>
-                    </ButtonZone>
-                    </>
-                )
-                }
-            </ModalView>
-        </ModalBackdrop>
-    )
+  return (
+    <ModalBackdrop>
+      <ModalView>
+        <CloseBtn onClick={handleCloseModal}>
+          <MdOutlineClose size="1.2rem" />
+        </CloseBtn>
 
-}
+        {type === ModalType.WELCOME ? (
+          <>
+            <ModalTitle>{title[0]}</ModalTitle>
+            <ButtonZone>
+              <Button type="primary" content="반가워요" onClick={handleCloseModal} />
+            </ButtonZone>
+          </>
+        ) : (
+          <>
+            <ModalTitle>{title[1]}</ModalTitle>
+            <ButtonZone>
+              <Button
+                type="cancel"
+                content="구독취소"
+                onClick={handleCloseModal}
+                width="calc(30%-0.5rem)"
+              />
+              <Button
+                type="basic"
+                content="닫기"
+                onClick={handleCloseModal}
+                width="calc(40%-0.5rem)"
+              />
+            </ButtonZone>
+          </>
+        )}
+      </ModalView>
+    </ModalBackdrop>
+  );
+};
 
 export default Modal;
 
@@ -65,7 +72,7 @@ const ModalBackdrop = tw.div`
     justify-center
     items-center
     bg-black/50
-`
+`;
 const ModalView = tw.div`
     w-1/3
     h-1/3
@@ -86,7 +93,7 @@ const ModalView = tw.div`
 const CloseBtn = tw.button`
     ml-auto
     bg-transparent
-`
+`;
 
 const ModalTitle = tw.div`
     text-xl
@@ -94,9 +101,9 @@ const ModalTitle = tw.div`
     mb-5
     font-medium
 
-`
+`;
 const ButtonZone = tw.div`
     flex
     gap-12
     justify-center
-`
+`;
